requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
discord.py-self==1.9.2
idna==3.4
multidict==6.0.3
requests==2.28.1
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2